60411 Corrigido percentual de repostas por opção nos rel. Mat.

+++ src/SME.SR.Application/Queries/Sondagem/MatematicaAditivoMultiplicativoConsolidado/ObterSondagemMatematicaAditivoMultiplicativoConsolidado/ObterSondagemMatAditMultiConsolidadoQueryHandler.cs
@@ -180,7 +180,7 @@ namespace SME.SR.Application
                                 PerguntaId = pergunta.Pergunta.Equals("Ideia") ? 1 : 2,
                                 Resposta = resposta.Resposta,
                                 AlunosQuantidade = alunosQuantidade,
-                                AlunosPercentual = (alunosQuantidade / request.QuantidadeTotalAlunos) * 100
+                                AlunosPercentual = ((double)alunosQuantidade / request.QuantidadeTotalAlunos) * 100
                             });;
                         }
                     }
